% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rf_to_raster.R
\name{apply_rf_to_raster}
\alias{apply_rf_to_raster}
\title{Function to apply the trained Random Forest model on new raster data}
\usage{
apply_rf_to_raster(rf_model, raster_path, output_raster_path)
}
\arguments{
\item{rf_model}{Trained Random Forest model (from "train_rf_model"-function)}

\item{raster_path}{Path to the raster file with reflectance values (.tif-file)}

\item{output_raster_path}{Path to save the output raster file with predicted values}

\item{application:}{Apply created Random Forest Model to new Raster Data to generate a Prediction-Map. Output can be used in GIS software to look at exact prediction values of each pixel.}
}
\description{
Function to apply the trained Random Forest model on new raster data
}
\examples{
stacked_file <- system.file("data", "08-01-25_enkelboom_stacked.tif", package = "InFieldR")
rf_model <- system.file("data", "my_model.rds", package = "InFieldR")
# data("rf_model", package = "InFieldR")   # loads a trained randomForest model

# Predict and plot the result
result <- apply_rf_to_raster(rf_model, stacked_file, tempdir())
}
