% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal_Sample_Size.R
\name{optimal_sample_size}
\alias{optimal_sample_size}
\title{Extracting the optimal sample size of a study site generated through cLHS and KL Divergence analysis}
\usage{
optimal_sample_size(
  df = Bands,
  nb = 10,
  cseq = seq(10, 500, 10),
  its = 10,
  output_folder
)
}
\arguments{
\item{df}{Bands (Covariate dataframe:used by cLHS or sampling; used by KL divergence to calculate distribution)}

\item{nb}{number of bins for created quantiles}

\item{cseq}{sequence of sampled points by cLHS}

\item{its}{iterations of sampled data}

\item{output_file}{file directory}
}
\value{
csv file with optimal sample size with XYZ parameters
}
\description{
Extracting the optimal sample size of a study site generated through cLHS and KL Divergence analysis
}
