% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_shp_on_rast.R
\name{raster_stack_mask}
\alias{raster_stack_mask}
\title{Stack and mask of individual raster bands (from Sentinel 2)}
\usage{
raster_stack_mask(raster_scene_folder, aoi, output_folder)
}
\arguments{
\item{raster_scene_folder:}{Folder containing the raster files with the bands that will be used for the regression model (either 10/20/60 m Resolution)}

\item{aoi:}{Area of interest (AOI) as a vector file.}

\item{output_folder:}{Folder to save the output file (output directory where file will be saved)}
}
\description{
Stack and mask of individual raster bands (from Sentinel 2)
}
