% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_shp_on_rast1.R
\name{stack_shp_on_rast}
\alias{stack_shp_on_rast}
\title{Stack shape file on raster data and mask}
\usage{
stack_shp_on_rast(shp_file, raster_file, output_folder)
}
\arguments{
\item{shp_file}{SHP file with in situ-measurments}

\item{raster_file}{Raster file with satellite scene from the closest date of the in-situ data acquisition}

\item{output_folder}{Folder to save the output file (output directory)}
}
\description{
Stack shape file on raster data and mask
}
\examples{
stack_shp_on_rast("in-situ_values.shp", "S1_scene_010125.tif", "output_folder")
}
