% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_rf_model2.R
\name{train_rf_model}
\alias{train_rf_model}
\title{Train Random Forest Regression Model}
\usage{
train_rf_model(
  regression_input_path,
  response_var,
  output_folder,
  ntree = 500,
  mtry = 3
)
}
\arguments{
\item{regression_input_path}{Path to the input file containing the regression data (Output file from "stack_shp_on_rast")}

\item{response_var}{Name of the in-situ data column (response-variable) in the regression data}

\item{output_folder}{Folder to save the trained model}

\item{ntree}{Number of trees in the Random Forest model}

\item{mtry}{Number of variables randomly sampled as candidates at each split}
}
\description{
Train Random Forest Regression Model
}
\examples{
train_rf_model("insitu_reflectance.txt", "Label", output_folder, ntree = 500, mtry = 3)
}
