% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rf_to_raster3.R
\name{apply_rf_to_raster}
\alias{apply_rf_to_raster}
\title{Apply created Random Forest Model to new/unseen Raster Data to generate a Prediction-Map}
\usage{
apply_rf_to_raster(rf_model, raster_path, output_raster_path)
}
\arguments{
\item{rf_model}{Trained Random Forest model}

\item{raster_path}{Path to the raster file with reflectance values (.tif)}

\item{output_raster_path}{Path to save the output raster file with predicted values}
}
\description{
Apply created Random Forest Model to new/unseen Raster Data to generate a Prediction-Map
}
