% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_rf_to_raster.R
\name{apply_rf_to_raster}
\alias{apply_rf_to_raster}
\title{Function to apply the trained Random Forest model on new raster data
Apply created Random Forest Model to new/unseen Raster Data to generate a Prediction-Map. Output can be used in GIS software to look at exact prediction values of each pixel.}
\usage{
apply_rf_to_raster(rf_model, raster_path, output_raster_path)
}
\arguments{
\item{rf_model}{Trained Random Forest model (from "train_rf_model"-function)}

\item{raster_path}{Path to the raster file with reflectance values (.tif-file)}

\item{output_raster_path}{Path to save the output raster file with predicted values}
}
\description{
Function to apply the trained Random Forest model on new raster data
Apply created Random Forest Model to new/unseen Raster Data to generate a Prediction-Map. Output can be used in GIS software to look at exact prediction values of each pixel.
}
\examples{
data("my_raster", package = "InFieldR")  # loads a SpatRaster
data("rf_model", package = "InFieldR")   # loads a trained randomForest model

# Predict and plot the result
result <- apply_rf_to_raster(rf_model, my_raster, tempdir())
terra::plot(result, main = "Predicted Raster from Random Forest")
}
