% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_shp_on_rast.R
\name{stack_shp_on_rast}
\alias{stack_shp_on_rast}
\title{Stack shape file on raster data and mask}
\source{
Copernicus hub and in-situ data from a field in Enkelboom, South Africa
}
\usage{
stack_shp_on_rast(shp_file, raster_file, output_folder)
}
\arguments{
\item{shp_file}{stacked SHP file (your own or the output from function "raster_stack_mask") with in-situ measurements}

\item{raster_file}{Raster file with satellite scene from the closest date of the in-situ data acquisition (Sentinel-scene from e.g. copernicus-hub)}

\item{output_folder}{Folder to save the output file (output directory)}
}
\description{
Stack shape file on raster data and mask
}
\examples{
data("shp_enkelboom", package = "InFieldR")
#print(shp_enkelboom)
data("raster_enkelboom", package = "InFieldR")
terra::unwrap(raster_enkelboom)
#print(raster_enkelboom)
if (exists("shp_enkelboom") && exists("raster_enkelboom")) {
  result <- stack_shp_on_rast(shp_enkelboom, raster_enkelboom, tempdir())
}

}
